Chore: Split the xdp_collector.go file into multiple files for better modularity
